requirements.txt
altgraph==0.17.2
certifi==2022.6.15
charset-normalizer==2.1.0
future==0.18.2
idna==3.3
importlib-metadata==4.12.0
Naked==0.1.31
pefile==2022.5.30
pycryptodome==3.15.0
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
typing-extensions==4.3.0
urllib3==1.26.10
zipp==3.8.1
